requirements.txt
Flask==1.1.2
ruamel.yaml==0.16.10
dataflows==0.0.71
pycurl==7.43.0.5
psycopg2-binary==2.8.5
Flask-HTTPAuth==3.3.0
